requirements.txt
pyyaml
gitpython
rich
